English: 1. Algorithm fix and update changes.
2.Description
The generation algorithm for each round of subkeys has been changed.
If you find any problems on the way to use this version, feel free to give me feedback.

3.Fix bugs

Chinese:

1.算法修复和更新更改

2.说明
每轮子密钥的生成算法已更改。
如果你在使用这个版本的途中发现任何问题，欢迎跟我反馈。

3.修复错误和丢失的文件

+++ include/CommonSecurity/CommonSecurity.hpp
@@ -232,7 +232,7 @@ namespace CommonSecurity
 			integer_a = static_cast<std::uint32_t>( LogarithmicTable[integer_a] );
 			integer_b = static_cast<std::uint32_t>( LogarithmicTable[integer_b] );
 
-			auto value = static_cast<std::int32_t>(integer_a + integer_b) % 255;
+			auto value = (integer_a + integer_b) % 255;
 
 			return static_cast<ByteType>( ExponentialTable[value] );
 		}
